Create and hook up Config struct
This struct contains the configuration for Zigup. A global object
`config` is initialized from command-line arguments and from reading in
a ZON file. Its ensured that command-line arguments always overwrite ZON
file values, and if neither are set then the default value is used.

A global (`zigup.config`) is used for storing the final configuration
because its pretty universal and gets used all over the place. Its
expected that this value gets initialized once, before we start doing
any serious work, and is never mutated again.

Some shuffling around of the parsing was needed to be done, notably the
check for the `run` command was moved after configuration is
initialized.

The `getInstallDir` and `makeZigPathLinkString` have been moved and
renamed as functions within Config, and functions using those instead
read from `zigup.config` directly. This has an added benefit of removing
some allocations.