Change default install path to $XDG_DATA_HOME/zigup on non-Windows systems
Fixes #177.  Also related to discussion in #146.

Windows path is not changed because currently on Windows the path is
derived from the location of the binary, so assuming Windows does not
allow two identcal paths, the most logical path name, `zigup`, is not
available.

I don't know about Windows conventions and don't have a Windows machine
available to be able to test my assumptions.